Implementando a lista de tarefas e tarefa

+++ src/main/java/br/com/tarefas/model/persistence/dao/ListaTarefaDAO.java
@@ -36,7 +36,7 @@ public class ListaTarefaDAO extends HibernateUtil{
 		} 
 	}
 	
-	public ListaTarefa findById(int id){
+	public ListaTarefa findById(long id){
 		try {
 			return (ListaTarefa) em.find(ListaTarefa.class,id);
 		} catch (Exception e) {
